requirements-prod.txt
# Production dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
email-validator>=2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Task Queue
celery==5.3.4
redis==5.0.1
APScheduler==3.10.4

# OpenAI (for voice)
openai==1.3.5

# Web scraping & automation
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0

# SAT Integration (SOAP/WSDL)
cfdiclient>=1.6.0
zeep>=4.2.1
psutil>=5.9.0
asyncpg>=0.29.0
lxml>=4.9.0
cryptography>=41.0.0

# OCR
pytesseract==0.3.10
pillow==10.1.0

# Utilities
python-dotenv==1.0.0
jinja2==3.1.2
numpy==1.24.3

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# CORS
fastapi-cors==0.0.6